chore: rename duplicate mock index file
