+++ src/__tests__/pages/pupils/lessons/[lessonSlug]/results/[attemptId]/printable.test.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { screen, waitFor } from "@testing-library/react";
 import { oakDefaultTheme, OakThemeProvider } from "@oaknational/oak-components";
 
@@ -8,7 +7,7 @@ import CanonicalResultsPage, {
   getStaticProps,
   InnerRender,
 } from "@/pages/pupils/lessons/[lessonSlug]/results/[attemptId]/printable";
-import curriculumApi2023 from "@/node-lib/curriculum-api-2023/__mocks__/index";
+import curriculumApi2023 from "@/node-lib/curriculum-api-2023/__mocks__/curriculumApi2023Mock";
 import { lessonBrowseDataFixture } from "@/node-lib/curriculum-api-2023/fixtures/lessonBrowseData.fixture";
 import { lessonContentFixture } from "@/node-lib/curriculum-api-2023/fixtures/lessonContent.fixture";
 import OakError from "@/errors/OakError";
