+++ src/__tests__/pages/pupils/years/[yearSlug]/subjects.test.tsx
@@ -2,7 +2,7 @@ import PupilSubjectListing, {
   getStaticPaths,
   getStaticProps,
 } from "@/pages/pupils/years/[yearSlug]/subjects";
-import * as curriculumApi2023 from "@/node-lib/curriculum-api-2023/__mocks__/index";
+import * as curriculumApi2023 from "@/node-lib/curriculum-api-2023/__mocks__/curriculumApi2023Mock";
 import { subjectBrowseDataFixture } from "@/node-lib/curriculum-api-2023/fixtures/subjectBrowseData.fixture";
 import renderWithProviders from "@/__tests__/__helpers__/renderWithProviders";
 import { topNavFixture } from "@/node-lib/curriculum-api-2023/fixtures/topNav.fixture";
