+++ src/node-lib/curriculum-api-2023/__mocks__/curriculumApi2023Mock.ts
