+++ src/pages-helpers/curriculum/downloads/getMvRefreshTime.test.ts
@@ -1,6 +1,6 @@
 import { getMvRefreshTime } from "./getMvRefreshTime";
 
-import * as curriculumApi2023 from "@/node-lib/curriculum-api-2023/__mocks__/index";
+import * as curriculumApi2023 from "@/node-lib/curriculum-api-2023/__mocks__/curriculumApi2023Mock";
 const mockErrorReporter = jest.fn();
 jest.mock("@/common-lib/error-reporter", () =>
   jest.fn(() => mockErrorReporter),
